[wip] working out how to read block file Signed-off-by: Owen Diehl <ow.diehl@gmail.com>
